changed typingusers.js
